kurang update product

+++ lib/entity/product.dart
@@ -1,7 +1,7 @@
 class Product {
   final int? id;
   final String? nama;
-  final int? harga;
+  final double? harga;
   final int? durasi;
 
   Product({this.nama, this.harga, this.durasi, this.id});
